+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/session/profile/ProfileModule.kt
@@ -42,4 +42,13 @@ internal abstract class ProfileModule {
 
     @Binds
     abstract fun bindGetProfileTask(task: DefaultGetProfileInfoTask): GetProfileInfoTask
+
+    @Binds
+    abstract fun bindRefreshUserThreePidsTask(task: DefaultRefreshUserThreePidsTask): RefreshUserThreePidsTask
+
+    @Binds
+    abstract fun bindBindThreePidsTask(task: DefaultBindThreePidsTask): BindThreePidsTask
+
+    @Binds
+    abstract fun bindUnbindThreePidsTask(task: DefaultUnbindThreePidsTask): UnbindThreePidsTask
 }